ajax通信のための準備
restcliantを用いて、データ通信の仕組みを理解した。
jQueryのslimを消して、minを入れた
Shift modelにfillableを追加
idのデータ型を変更した

+++ database/migrations/2020_09_15_130134_create_shifts_table.php
@@ -14,7 +14,7 @@ class CreateShiftsTable extends Migration
     public function up()
     {
         Schema::create('shifts', function (Blueprint $table) {
-            $table->INTEGER('id');
+            $table->bigIncrements('id');
             $table->INTEGER('user_id');
             $table->DATE('shift_date');
             $table->DATETIME('shift_start');
